+++ ntlm3/HTTPNtlmAuthHandler.py
@@ -17,8 +17,10 @@ from six.moves.http_client import HTTPConnection, HTTPSConnection
 import socket
 import re
 
+from ntlm3.constants import NegotiateFlags
 from . import ntlm
 
+# TODO: Is this file needed, this is very much what requests-ntlm does and should be brought there. This hasn't been updated ot support NTLMv2
 
 class AbstractNtlmAuthHandler:
     def __init__(self, password_mgr=None, debuglevel=0):
@@ -46,7 +48,7 @@ class AbstractNtlmAuthHandler:
             if len(user_parts) == 1:
                 UserName = user_parts[0]
                 DomainName = ''
-                type1_flags = ntlm.NTLM_TYPE1_FLAGS & ~ntlm.NTLM_NegotiateOemDomainSupplied
+                type1_flags = ntlm.NTLM_TYPE1_FLAGS
             else:
                 DomainName = user_parts[0].upper()
                 UserName = user_parts[1]
@@ -90,7 +92,8 @@ class AbstractNtlmAuthHandler:
 
             r.begin()
 
-            r._safe_read(int(r.getheader('content-length')))
+            a = r.getheader('Content-Length')
+            #r._safe_read(int(r.getheader('Content-Length')))
             if r.getheader('set-cookie'):
                 # this is important for some web applications that store authentication-related info in cookies (it took a long time to figure out)
                 headers['Cookie'] = r.getheader('set-cookie')