+++ test/integration/test_ntlm_auth_handler_edge_cases.py
@@ -1,4 +1,4 @@
-import unittest
+import unittest2 as unittest
 import pytest
 from httpretty import HTTPretty, httprettified, Response