+++ CHANGES.md
@@ -17,3 +17,21 @@ Initial public offering.
 * Bugfix for windows (thanks to @rbcarson for the help)
 * Added Appveyor continous integration testing on Windows to avoid these problems in future
 
+1.1.0 (Aug 5, 2016)
+-------------------
+
+* Added support for Python 3.5
+* Added requirement for cryptography so we can calculate RC4 values for EncryptedRandomSessionKey and signing and sealing (can we remove this dependency?)
+* Major rewrite of how python-ntlm3 handles authentication
+* Added support for NTLMv2 auth and fixed up some older auth methods
+* Moved code to separate classes to help cleanup the code
+* Added support for channel_bindings (CBT) when supplying a certificate hash
+* Added support for MIC data for authenticate messages
+* Added support for signing and sealing of messages. Actual message structure is done outside of python-ntlm3 in the app level but this can still encrypt whatever message is sent through.
+* Removed some methods that weren't being used at all (most were starting to implement these features above but weren't there)
+* More comments on each methods relating back to the MS-NLMP document pack on NTLM authentication for easier maintenance
+* Created target_info.py to handle AV_PAIRS and putting it in the target info
+* Renaming of some variables to match more closely with the Microsoft documentation, makes it easier to understand what is happening
+* Rewriting of tests to accommodate these new changes and to cover the new cases
+* Added better itegration test in Appveyor to test a connection to a local IIS server configured to only allow NTLM authentication
+* The methods `create_NTLM_NEGOTIATE_MESSAGE`, `parse_NTLM_CHALLENGE_MESSAGE`, `create_NTLM_AUTHENTICATE_MESSAGE` will no longer be supported in future version. They do not support NTLMv2 auth and are only left for compatibility